New ListBoxItem and UnitBoxItem controls.
The ListBoxItem is a simple drop-down listbox, similar in use to other
FormItems. To populate the possible choices in the listbox call
  ListBoxItem.setChoices()
The UnitBoxItem is a text field that has an associated unit field. To
populate the possible units, call UnitBoxItem.setChoices().
The UnitBoxItem is really two FormItems in one, therefore when its added
to a form the unit field needs to be added separately from the text
field. E.g.:
  form.setField(item, item.getUnitItem());

Additionally, GWT unit tests for the above controls are included and a
GWTTestSuite is added to speed up the running of the unit tests.